New base class `PersistentStorageException`
There needs to be some abstract exception type that can be used in
the description of the contract defined by `IPersistentStorage`. For
instance, this contract should be able to specify that `GetItemAsync`
will throw (as opposed to e.g. returning a null reference) when given
an ID of an item that is not present in the storage.